OCR_backend/requirements.txt
transformers==4.48.3
qwen-vl-utils==0.0.10
torch==2.6.0
pillow==11.1.0
nltk==3.9.1
contractions==0.1.73
torchvision==0.21.0
uvicorn==0.34.1
fastapi==0.110.1
python-multipart==0.0.9
scikit-learn==1.6.1
sentence_transformers==3.4.1
sumy==0.11.0